+add table component vue , *change ape.vue, +add tests for userservice

+++ chat/chat/src/main/java/wildau/web/live/chat/Service/UserService.java
@@ -1,6 +1,6 @@
 package wildau.web.live.chat.Service;
 
-import org.apache.coyote.BadRequestException;
+import org.apache.coyote.BadRequestException ;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 import wildau.web.live.chat.api.model.UserEntity;
+import wildau.web.live.chat.exceptions.UserEntityNotFoundException;
 import wildau.web.live.chat.repository.UserRepository;
 @Service
 public class UserService {
@@ -27,24 +28,34 @@ public class UserService {
         return userRepository.findById(userId);
     }
     
-
-
     public UserEntity updateUser(UserEntity user) {
-       
+
+         if(!userRepository.existsById(user.getId())){
+            throw new UserEntityNotFoundException("User with id "+ user.getId()+" does not exsists => Update failed");
+        }
         return userRepository.save(user);
     }
+
     public void addUser(UserEntity newUser) throws BadRequestException {
 
-        Boolean existsEmail = userRepository.selectExistsEmail(newUser.getEmail());
-        if(existsEmail){
-            throw new BadRequestException("Email" + newUser.getEmail()+"exsitiert bereits");
+        Boolean excepted = userRepository.selectExistsEmail(newUser.getEmail());
+        if(excepted){
+            throw new BadRequestException("Email " + newUser.getEmail()+" is assigned");
         }
-        
-        userRepository.save(newUser);
-       
+        userRepository.save(newUser); 
     }
 
     public void deleteUser(Long userId) {
+        if(!userRepository.existsById(userId))
+        {
+            throw new UserEntityNotFoundException("User with id "+ userId+" does not exsists => Delete failed");
+        }
         userRepository.deleteById(userId);
     }
+    public List<UserEntity> getUserByNames(String keyword){
+        if(keyword !=null) {
+         return userRepository.search(keyword);
+        }
+        return userRepository.findAll();
+     }
 }
